hierarchy re-design, feat, fix, refactor, code documentation
1. Code is divided into modules: "model", "plotting", "static" (fonts, images, styles), "ui", "data_processing" (database).
2. Pie-chart is now based on ranges instead of specific values (so far only works correctly for the "age" column).
3. Fixed a bug where a large csv file with Cyrillic characters in the string was not imported due to encoding mismatch.
4. Code refactoring.
5. Started basic documentation of the code in Russian.
6. Changed the application icon.
---
1. Код разделён на модули: модель, начертание графиков, статика (шрифты, изображения, стили), ui, манипуляции с данными (БД).
2. Круговая диаграмма теперь основана на диапазонах, а не на конкретных значениях (пока что корректно работает лишь для столбца "возраст").
3. Исправлена ошибка, при которой из-за несоответствия кодировки не импортировался большой файл-csv с кириллицей в строка.
4. Рефакторинг кода.
5. Начата базовая документация кода на русском языке.
6. Изменена иконка приложения.
