+++ main.py
@@ -1,0 +1,9 @@
+import sys
+from PyQt6.QtWidgets import QApplication
+from ui.main_window import MainWindow
+
+if __name__ == "__main__":
+    app = QApplication(sys.argv)
+    window = MainWindow()
+    window.showMaximized()
+    sys.exit(app.exec())
